Simplify bootstrap
